% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/affine.R
\name{buildAffine}
\alias{buildAffine}
\title{Build an affine matrix up from its constituent transformations}
\usage{
buildAffine(translation = c(0, 0, 0), scales = c(1, 1, 1), skews = c(0, 0,
  0), angles = c(0, 0, 0), source = NULL, target = NULL)
}
\arguments{
\item{translation}{Translations along each axis, in \code{\link{pixunits}}
units. May also be a list, such as that produced by
\code{\link{decomposeAffine}}, with elements for translation, scales,
skews and angles.}

\item{scales}{Scale factors along each axis.}

\item{skews}{Skews in the XY, XZ and YZ planes.}

\item{angles}{Roll, pitch and yaw rotation angles, in radians.}

\item{source}{The source image for the transformation.}

\item{target}{The target image for the transformation. Can be the same as
\code{source}, if appropriate.}
}
\value{
A 4x4 affine matrix representing the composite transformation.
}
\description{
This function does the opposite to \code{\link{decomposeAffine}}, building
up an affine matrix from its components. It can be useful for testing.
}
\author{
Jon Clayden <code@clayden.org>
}
\seealso{
\code{\link{decomposeAffine}}, \code{\link{isAffine}}
}

