% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transform.R
\name{halfTransform}
\alias{halfTransform}
\title{Calculate a half transformation}
\usage{
halfTransform(transform)
}
\arguments{
\item{transform}{A transform, possibly obtained from \code{\link{forward}}
or \code{\link{reverse}}.}
}
\value{
The half-way transform, in a similar format to \code{transform}.
}
\description{
This function calculates the half-way transformation corresponding to its
argument. Applying this transformation results in points or images in a
space halfway between the original source and target images, which can be a
useful common space in some applications.
}
\author{
Jon Clayden <code@clayden.org>
}
\seealso{
\code{\link{niftyregLinear}}, \code{\link{niftyregNonlinear}}
}

