% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/image.R
\name{dim.internalImage}
\alias{as.array.internalImage}
\alias{dim.internalImage}
\alias{internalImage}
\title{Internal images}
\usage{
\method{dim}{internalImage}(x)

\method{as.array}{internalImage}(x, ...)
}
\arguments{
\item{x}{An \code{"internalImage"} object.}

\item{...}{Additional parameters to methods. Currently unused.}
}
\description{
An internal image is a simple R object with a few attributes including a
pointer to an internal C structure, which contains the full image data. They
are used in the package for efficiency, but can be converted to a normal
R array using the \code{as.array} method.
}
\author{
Jon Clayden <code@clayden.org>
}

