% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\name{invertAffine}
\alias{invertAffine}
\title{Invert an affine matrix}
\usage{
invertAffine(affine)
}
\arguments{
\item{affine}{An existing 4x4 affine matrix.}
}
\value{
The inverted affine matrix.
}
\description{
This function is used to invert an affine matrix. It is a wrapper around
\code{\link{solve}}, which additionally sets appropriate attributes.
}
\examples{
affine <- readAffine(system.file("extdata","affine.txt",package="RNiftyReg"))
print(affine)
print(invertAffine(affine))

}
\author{
Jon Clayden <code@clayden.org>
}
\seealso{
\code{\link{solve}}
}

