% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{jacobian}
\alias{jacobian}
\title{Extract a Jacobian determinant map}
\usage{
jacobian(x)
}
\arguments{
\item{x}{An R object, probably a deformation field.}
}
\description{
This function extracts the Jacobian determinant map associated with a
deformation field.
}
\author{
Jon Clayden <code@clayden.org>
}
\seealso{
\code{\link{deformationField}}
}

