% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/image.R
\name{pixdim}
\alias{pixdim}
\alias{pixunits}
\title{Pixel dimensions and units}
\usage{
pixdim(object)

pixunits(object)
}
\arguments{
\item{object}{An R object, generally an image.}
}
\value{
\code{pixdim} returns a numeric vector of pixel dimensions.
  \code{pixunits} returns a character vector of length up to two, giving the
  spatial and temporal unit names.
}
\description{
These functions return the \code{"pixdim"} and \code{"pixunits"} attributes
of their arguments. These represent the physical step size between pixel or
voxel centre points, and the spatial and temporal units that they are given
in. The former defaults to 1 in each dimension, if there is no attribute.
}
\author{
Jon Clayden <code@clayden.org>
}

