% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nifti.R
\name{readNifti}
\alias{readNifti}
\title{Read a NIfTI-1 format file}
\usage{
readNifti(file, internal = FALSE)
}
\arguments{
\item{file}{A character vector of file names.}

\item{internal}{Logical value. If \code{FALSE} (the default), an array
containing the image pixel or voxel values will be returned. If
\code{TRUE}, the return value will be an object of class
\code{"internalImage"}, which contains only minimal metadata about the
image. Either way, the return value has an attribute which points to a
C data structure containing the full image.}
}
\value{
An array or \code{"internalImage"} object.
}
\description{
This function reads one or more NIfTI-1 files into R, using the standard
NIfTI-1 C library.
}
\author{
Jon Clayden <code@clayden.org>
}
\references{
The NIfTI-1 standard (\url{http://nifti.nimh.nih.gov/nifti-1}).
}

