% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{saveTransform}
\alias{saveTransform}
\alias{loadTransform}
\title{Save and load transform objects}
\usage{
saveTransform(transform, file)

loadTransform(file)
}
\arguments{
\item{transform}{A transform, possibly obtained from \code{\link{forward}}
or \code{\link{reverse}}.}

\item{file}{The filename to save to, or load from.}
}
\value{
\code{loadTransform} returns a deserialised transform object.
}
\description{
These objects save a full transformation object, including source and target
image metadata, to a self-contained RDS file, or load it back from such a
file. This is currently only possible for linear transforms.
}
\seealso{
\code{\link{writeAffine}}, \code{\link{readAffine}}
}
\author{
Jon Clayden <code@clayden.org>
}
