% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transform.R
\name{composeTransforms}
\alias{composeTransforms}
\title{Compose two transformations}
\usage{
composeTransforms(transform1, transform2)
}
\arguments{
\item{transform1,}{transform2 Affine or nonlinear transforms, possibly
obtained from \code{\link{forward}} or \code{\link{reverse}}.}
}
\value{
The composed transform. If both \code{transform1} and
  \code{transform2} are affines then the result will also be an affine;
  otherwise it will be a deformation field.
}
\description{
Compute the composition of two transforms, the single transform that
combines their effects in order.
}
\author{
Jon Clayden <code@clayden.org>
}
\seealso{
\code{\link{niftyregLinear}}, \code{\link{niftyregNonlinear}},
  \code{\link{deformationField}}
}

