% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{composeTransforms}
\alias{composeTransforms}
\title{Compose transformations}
\usage{
composeTransforms(...)
}
\arguments{
\item{...}{Affine or nonlinear transforms, possibly obtained from
\code{\link{forward}} or \code{\link{reverse}}.}
}
\value{
The composed transform. If all arguments are affines then the result
  will also be an affine; otherwise it will be a deformation field.
}
\description{
Compute the composition of two or more transforms, the single transform that
combines their effects in order.
}
\author{
Jon Clayden <code@clayden.org>
}
\seealso{
\code{\link{niftyreg.linear}}, \code{\link{niftyreg.nonlinear}},
  \code{\link{deformationField}}
}

