% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\name{writeAffine}
\alias{writeAffine}
\title{Write an affine matrix to a file}
\usage{
writeAffine(affine, fileName, comments = TRUE)
}
\arguments{
\item{affine}{A 4x4 affine matrix.}

\item{fileName}{A string giving the file name to write the matrix to.}

\item{comments}{Logical value: if \code{TRUE} comments are written to the
file in lines beginning with \code{#}.}
}
\description{
This function is used to write a 4x4 numeric matrix representing an affine
transformation to a file. A comment is also (optionally) written, which
specifies the matrix as using the NiftyReg convention, for the benefit of
\code{\link{readAffine}}.
}
\seealso{
\code{\link{write.table}}, \code{\link{readAffine}}
}
\author{
Jon Clayden <code@clayden.org>
}
