% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nifti.R
\name{writeNifti}
\alias{writeNifti}
\title{Write a NIfTI-1 format file}
\usage{
writeNifti(image, file)
}
\arguments{
\item{image}{An image, in any acceptable form (see \code{\link{isImage}}).}

\item{file}{A character string containing a file name.}
}
\description{
This function writes an image to NIfTI-1 format, using the standard NIfTI-1
C library.
}
\author{
Jon Clayden <code@clayden.org>
}
\references{
The NIfTI-1 standard (\url{http://nifti.nimh.nih.gov/nifti-1}).
}

